Add metadata validation via json schema (#129)
-Schema files must include a .schema suffix so as to distinguish them
from other metadata json files and must reside in the same directory
as associated .json files.

-Currently, the schema_name entry is not required in the metadata
json files. If present, the schema will be validated.

-Although 'name' is required by the Metadata class, it is not required in
the schema. This is because the current implementation derives the name
from the json filename.

-A singleton "cache" of schemas is maintained. Epochs are provided
by the caller (File, DB, etc.) and, in most cases, it prevents the need from
reading the schema into memory each time a metadata file is read. 
Not sure how useful this might be although the savings might add up 
if there are lots of metadata files of a given schema.

Fixes: #109