Update pipeline submission code show expected dialogs (#146)
The current submission code throws an error in the console before
getting to the dialog code. This causes a submission's success or
failure to not surface to the user as expected.

This change updates both the pipeline and notebook submission
code to surface a success or failure dialog on submission.

This change can handle success and error cases, but due to issues
in the handler code does not display the reason for errors. This has 
been raised in #145

Fixes #112