[HOTFIX] Revert to link dev version of lab extension
There is a bug in JupyterLab where if a extension with
same version gets updated it's not built by lab, and
thus the staging area does not gets updated causing
the runtime to see old/stale version of the extension
code.

For dev scenarios, link is a workaround, but cause
multiple build requests if code on the directory changes.