Update pipeline add to parse notebook for env vars
Parses a notebook when adding it to a pipeline like the notebook
submit button does. Then prepopulates the vars comm prop with the
expected env vars.

This adds a small delay to the add step since in order to grab the
notebook content a notebook widget is created then disposed of.