Add support to install and remove runtimes
These changes extend the juptyer-runtime application to support
installation and removal of runtime metadata.  For installs, it
supports adding `kfp` runtime metadata and provides a class that
will contain `airflow` runtime support.  For removes, only the
name of the metadata is required.

Also included in this PR are tests of install (create), remove,
and list.  I also added sorting to the `jupyter-runtime list`
results, primarily for test purposes, but it also makes the output
more organized.

Fixes #199