Internalize system-owned metadata schema (#206)
Move metadata schema into the package under metadata/schemas.  
This is because this metadata is closely tied to the implementation.  
Eventually, we'll move validation close to the processor implementations
but has been left in its current location for now.

Removed the staleness logic from SchemaManager since schemas
should not change except across release boundaries.  As a result, 
we may want to consider storing version information in both the 
schema and instance data.

Fixes ##203