Include invalid runtime metadata when listing runtimes
`jupyter runtime list` will now include invalid runtime metadata in
its output.  The invalid metadata will be marked with an entry of
**INVALID** (<reason>) where <reason> will likely be the exception
class name.

Users wishing to not include invalid metadata can use the new option
`--valid-only`, in which case the invalid metadata will not be included.
However, because logging goes to stderr, users will typically see the
log message that includes the details of the invalid metadata.

Also fixed a bug in `jupyter runtime list --json` which was not handling
the JSON correctly.

Fixes: #213