Add support for pipeline node icons
Currently all pipleine nodes use the jupyter logo as thier icon.

With the update to LabIcons in 2.0 it is now easy to support the use of
file sprecific icons for each node in a pipeline. Currently this
implementation is hard-coded to the notebookIcon since only notebooks
are currently supported for nodes