Add frontend to allow for exporting pipelines
First pass at enabling exporting pipelines as different file types
such as tar, tar.gz, zip, etc. Currently, the behavior is that you
press an export button in the canvas, it opens a dialog that
prompts for the name of the pipeline, the runtime to use, and the
filetype to export as. When you click "Ok" it opens a new dialog
that says "Submitting pipeline..." (Need to change) then when the
compilation finishes, it says  that the pipeline was submitted
successfully (also need to change). Might need to change the
implementation to not use the same functions as pipeline submissions.
Also want to look into the possibility of having an "Export as..."
option in the right-click menu for a pipeline file.