Adds check box to overwrite current file
Changes the way that the pipeline application confirms whether an
exported file should overwrite an existing exported file by adding
a check box to ask user if they want to overwrite an existing file,
and raising an error if the box is unchecked but the export would
overwrite a file. Also changes the status code of the post request
from 200 to 201 to be in accordance with status code standards.
Note: another PR is going to create more modular handling of status
codes in packages/application/src/submission.tsx.