Properly include dependencies recursively (#476)
There were a couple of issues when supporting dependencies
recursively. First, the property was not being saved/propagated.
After that, because the code was including the notebook as one
of the files on the dependency array, the algorithm needed to be
updated to include subdirs when processing the dependencies.
Note, there was no problem if recursive was true and no file
dependencies were passed.