Merge notebok-scheduler into pipeline-editor (#505)
Phase 1 of merging the notebook-scheduler and pipeline-editor
extensions. This first step moves all functionality of
notebook-scheduler into pipeline-editor "as is"

Includes:
- moves non-index files into pipeline-editor from notebook-scheduler
- copies code in index.ts and index.css between packages
- updates build files with missing dependencies and imports
- updates docs to remove notebook-scheduler
- deletes notebook-scheduler from packages

This will be followed up with more updates which will address
updating and moving utils and then refactoring pipeline-editor
to unify the code copied in this PR and further modularize it.