Return No Content on DELETE, set Location header on POST
This change no longer returns the deleted resource (and 200) when
deleting metadata instance and instead returns 204 (No Content).

Also, while researching REST behaviors, I found that we should be
setting the Location header when creating instances.  The value of
which is the endpoint that would be used to fetch that new content:
`/api/metadata/{namespace}/{resource}`.

Resolves #567