Add a make release target
Added a make release target that creates dists for the extensions
and includes them in the elyra wheel file for publishing.

Without this the published wheel would not install the front end.

This is specifically not for local builds since the extensions
installed via the wheel file will use the published application
and ui-components libraries.

Fixes #593