Fix bad JSON file handling
The error handling relative to JSON load issues was side-effected by
the recent refactor.  However, with these changes we now DO include
JSON load issues in the results when `include-invalid` is requested.

In such cases, name, schema_name, resource and reason are set on the
returned instance, although schema_name is set to '{unknown}' since
we cannot make that determination.