Use config to determine metadata storage class
The class to use for MetadataStore when creating a MetadataManager
instance is now pulled from the metadata_store_class trait, which
requires a class of MetadataStore and defaults to FileMetadataStore.

Differing storage classes can still be associated during MetadataManager
creation by using parameter `metadata_store_class = class-name`.

A `MockMetadataStore` has been created and the metadata manager and
metadata store tests now run against both it and FileMetadataStore
for testing.  This helps keep file-based items out of the metadata
manager logic and can serve as an example of a non-file based store
manager.