Enable CLI JSON output for script consumption (#746)
This change allows for `elyra-metadata list <namespace> --json` output to
be piped to JSON-aware tools for consumption.  It allows for usage w/ and
w/o the `--valid-only` flag where invalid entries will have a `reason`
with a value entry.

Errors still appear on stderr (as expected) but stdout can be piped or
redirected.