Add activate and remove restorer
Added an activate call to open the new sidebar tab once it's added.

Removed the restore because the widget doesn't exist on a reload
until the history button is clicked. Adding the widget earlier than
that is not an option since once a sidebar tab is added it cant be
hidden or closed, only programmitcally deleted