Add initial support for supernodes
Supernodes are now supported at a basic level.  The UI has been minimally
modified to enable their creation.  However, we'll likely want to enhance
the use of supernodes.

The pipeline submitted to the backend runtime consists of a flattened
pipeline where disjoint pipeline definitions can be executed in parallel.
The presence of a supernode essentially equates to the flattening of its
corresponding pipeline reconnected to any external input/output nodes.