Infer value of bool secure parameter from endpoint scheme (#774)
This change removes the need for maintaining an optional cos_secure field
in the metadata instance by inferring its boolean value from the scheme of
the corresponding cos_endpoint value (which is required).

Also updated the docs accordingly. Since the example, recommended setting
the previous cos_secure to True, the endpoint in the example is now prefixed
with a scheme of https.