Use get_event_loop instead of get_running_loop for py3.6
Although the docs recommend using get_running_loop(), it only exists
in Python 3.7.  This changes that to use the get_event_loop() method
which is essentially the same although the latter will create an
event loop if one does not exist.  Since we're running in a tornado
server, we're (virtually) guaranteed to have an event loop.