Add support to run pipelines in-place locally (#845)
Add support to run pipelines in-place in your local environment
without requiring availability/connectivity to a remote platform.

In this scenario, it is assumed that the pipeline is composed of
notebooks that already run in your local environment, and the pipeline
is used merely to identify the proper execution order. Note that we do
enable overriding the environment variables from the node property. 

Other things like isolation, running on a different environment (e.g. docker image) 
and all these additional capabilities are out of the scope of this mode and only
supported on an actual remote platform such as KFP.

Co-authored-by: Kevin Bates <kevin.bates@us.ibm.com>