Update pipeline editor to use relative paths (#861)
Currently, nodes in a pipeline save their paths relative to the
jupyter lab workspace. This updates them to use paths relative to
the pipeline file instead. This enables moving a pipeline "project",
such as moving the directory a pipeline's contents are in, checking
out a project from git, and starting lab from a different workspace.

A node's path is converted to be relative to the pipeline upon add
and that relative path is what's stored in the pipeline file. The path
is converted back to be relative to the workspace prior to being
submitted to the export or schedule API, or when a node is "opened".

Note that this change is designed to work in conjunction with
validation of the node files and the ability to change node file
locations.

Co-authored-by: Kevin Bates <kevin.bates@us.ibm.com>