Fix missing feild value in Metadata Editor
The latest release of JupyterLab updated to the latest release
of blueprintjs core, which included an async bugfix for
`InputGroup` which broke how our async update interacts with fields.
Causing any already rendered `InputGroup` to not update it value
on an async call to `uodate()` such as the one we do after calling
the metadata api.

I solved this by making the rendering of the display_name field
conditional on the existance of an inital value, just like we
already do with the other fields.

Fixes #883