Improve subprocess error handling (#955)
This change improves the messages that users receive
when encountering issues during node operation failures.

For both notebook and python processes, the capture_output
parameter has been enabled which results in the raised error
object instance representing the error in its stderr property. 
This property value is then decoded to unescape newlines - 
and stderr becomes part of the message that hereto now only
appeared in the server log file - as discussed on #951.

For notebook operations, it also turns off the progress bar 
output of the papermill command that had been captured
in the log file and was included in the stderr property following
the previous change.