Remove use of capture_output - not available in py 3.6 (#981)
This is essentially a rollback of #955. When making that change I failed
to realize that the subprocess run argument capture_output was only
available in Python versions >= 3.7. I happened to run on Python 3.6
and got an error. This also implies we need to fix up our local processor
tests since we should be able to run something locally.