Ensure local notebook nodes use local env (#1061)
This change mirrors that of #1047 but for notebooks. Also,
the local environment should be copied so as to not be 
side-affected by updates.

Updated the tests to include a pip install and import of the installed
module to ensure its completion. If there are other lighter and more obscure
packages than pygithub, suggestions are welcome. The primary requirement
is that it can't exist prior to CI.