Renames docker-image target to container-image (#1141)
Renames the existing Elyra make docker-image target to container-image
(so we start using more general terminology). 

It also adds a publish-container-image target, which pushes the Elyra
container image to Docker Hub and Quay.io. Note that publish is a privileged
task and only users with the appropriate access to the target project on 
hub.docker.com and quay.io can perform it by default after they've
logged in to the registries.