Update docs for clarity around creating KfDef YAML
In the documentation, after clicking 'Create KfDef' the user is taken
to a screen where they're shown the form view and given the option for YAML.
This change adds clarity, as on a quick glance it would appear the YAML view
is not there. Language was added to make this more obvious.