Update remaining notebook imports to jupyter_server (#1471)
Updated a couple more instances of notebook import statements
to reflect the jupyter_server location. The problem was because
we use GatewayClient from the "running server" to determine whether
a gateway server is configured or not and, if so, the configuration
settings to use (URL, timeouts, etc.). Since this is a singleton and
the reference is from the notebook package, yet jupyter_server is running, 
the notebook-relative singleton will not reflect the current configuration.