Refactor script processors, include brief detail on generic errors (#1485)
It introduces a new base class named ScriptOperationProcessor
from which the existingPythonScriptOperationProcessor and
RScriptOperationProcessor classes now derive. This base class
contains 90% of the applicable code with the subclasses providing
their name and argument vectors.

It introduces a log_and_raise() method on the base 
FIleOperationProcessor class that is available to all file-based
operations. Building on the work done in #1411, this method
checks the length of the error message and truncates it
to around the max (80), replacing overflow with ellipses (...).

Adds a test that removes the kernel metadata from a notebook
node and ensures the appropriate error is raised.