Add elyra-pipeline cli tool (#1246)
The purpose of this CLI is to be able to run pipelines locally
and to submit them to an external runtime outside the context
of JupyterLab.

The proposed user experience for the initial version is to introduce
two commands run and submit.

The run command would run the pipeline locally, in the context of CLI,
and there will be no support for runtime and runtime-config parameters.

Example of running a pipeline locally:
elyra-pipeline run this/is/an/example.pipeline

The submit would submit the pipeline to a supported remote runtime
such as KFP or Airflow. In this context, the runtime-config parameters
will point to the necessary configuration to submit the pipeline to the
remote runtime.

Example of submitting a pipeline to be run in the context of the Jupyter server:
elyra-pipeline submit this/is/an/example.pipeline --runtime-config kfp-shared-tekton

Co-authored-by: Luciano Resende <lresende@us.ibm.com>