Fix icon related console warnings and errors
IconUtil.colorize leverages a jupyterlab function that doesn't check
for duplicates before creating a new LabIcon, this adds caching to
prevent the duplicate LabIcon warning.

class is an invalid attribute when using react to render, but
className is an invalid attribute for svg. Since we don't reference
the class of our svg anyway I removed it to get rid of the error