Fix migration bugs
Fixes two bugs with the migration: multiple dialogs opening and
sometimes setting the runtime image property of the nodes to the
display name instead of the value.

The first bug was partial fixed in #2045 but the code setting the
flag to true was lost during revision.

The second bug described in #2208 was due to a mistake I made in #1860
by using  the transient copy of the pipeline used by the GUI instead of
the persisted content when running the migration.
This caused a race condition when the API call to get the runtime
images returned and triggered the changeHandler to update to the
display names before the user clicked the migrate button in the dialog.

Fixes #2208