Text Viewer Widget Initial POC
As an initial POC I've created the TextViewer as a widget within
the pipeline editor widget package. In a final version it would be
a separate package. The POC also includes a button the pipeline
toolbar that opens the viewer with the pipeline file contents as
an example of how it works.

Frontend POC for https://github.com/elyra-ai/elyra/issues/2549