Develop Elyra against the Jupyterlab source repo
Adds documentation and some additions to the build scripts to allow
Elyra developers to develope Elyra against a local build of
Jupyterlab.

Updates include:
- Updated documentation with step by step instructions on how to
step up the devlopment environment
- Updating our `page_config.json` to the new format and location.
We were previously using a deprecated format and location
uncompatable with this feature.
- Added a way to run `clean-jupyterlab.sh` without installing a new
version of jupyterlab
- Added new npm targets and make targets to leverage a yarn link
version of `@jupyterlab/builder`. This was added to the existing
support for `make dev-link`, but the order of make targets had to
be rearranged to make sure lerna was installed before calling it.
- Made a small backwards compatable change to fix a type conflict
found when running against jupyterlab 4.0.0 alpha

These changes are blocked by
https://github.com/jupyterlab/jupyterlab/pull/12533
without that change you also need to update the version of
`@jupyterlab/builder` in every package.json to prevent an error