Prepend baseUrl to icon source
Currently we only prepend `/` to the icon path when using icons
hosted on the server. This causes the icons to not be found when
the server has base_url set in in it's config.

To address this I've prepended baseUrl to the icon path instead as
we also do when making API calls to the server.

Fixes #2723