Do not persist the jupyter baseUrl in pipelines
The changes in #2728 cause the pipeline-editor to persist the full
url for a node's image. This could cause issues with surfacing PI,
as such I've updated that fix to instead use the full URL base,
which will behave the same as before the fix, but including the
URL path from the baseUrl if one exists.

Note that if a baseUrl path conatains PI that would still be
persisted, but puting PI in a URL path is considered inappropriate
handling of PI by the server owner, not Elyra.

Fixes #2773