Install jupyterlab during normal builds
In #2709 I removed jupyterlab from the build_requires.txt so that
the local versaion of jupyterlab could be used if installed. This
has caused frustration in users, so I've added the dependency back
and update the make dev-install target to not use the dependency
file at all, instead using a new make target to install the
required dependency.